# api/swagger.yaml
basePath: /beetle
definitions:
...
  controller.JSONResult:
    type: object
  controller.MigrateInfraRequest:
    properties:
      description:
        type: string
      status:
        type: string
      targetSecurityGroupList:
        items:
          $ref: '#/definitions/model.TbSecurityGroupReq'
        type: array
      targetVNet:
        $ref: '#/definitions/model.TbVNetReq'
      targetVmInfra:
        $ref: '#/definitions/model.TbMciReq'
      targetVmOsImageList:
        items:
          $ref: '#/definitions/model.TbImageInfo'
        type: array
      targetVmSpecList:
        items:
          $ref: '#/definitions/model.TbSpecInfo'
        type: array
    type: object
  controller.MigrateInfraResponse:
    type: object
...
  github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response:
    properties:
      details:
        example: Any details
        type: string
      list:
        items: {}
        type: array
      object:
        additionalProperties: true
        type: object
      success:
        example: true
        type: boolean
      text:
        example: Any text
        type: string
    type: object
...
  migration.MciInfoList:
    properties:
      mci:
        items:
          $ref: '#/definitions/model.MciStatusInfo'
        type: array
    type: object
...
  resource.RestGetAllSecurityGroupResponse:
    properties:
      securityGroup:
        items:
          $ref: '#/definitions/model.TbSecurityGroupInfo'
        type: array
    type: object
...
host: localhost:8056
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cm-beetle/issues/new/choose
  description: CM-Beetle REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Beetle REST API
  version: latest
paths:
...
  /migration/ns/{nsId}/mci:
    get:
      consumes:
      - application/json
      description: Get the migrated multi-cloud infrastructure (MCI)
      operationId: ListInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: status
        description: Option for getting the migrated multi-cloud infrastructure
        enum:
        - status
        - id
        in: query
        name: option
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The info list of the migrated multi-cloud infrastructure (MCI)
          schema:
            $ref: '#/definitions/migration.MciInfoList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Get the migrated multi-cloud infrastructure (MCI)
      tags:
      - '[Migration] Infrastructure'
    post:
      consumes:
      - application/json
      description: Migrate an infrastructure to the multi-cloud infrastructure (MCI)
        with defaults.
      operationId: MigrateInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Specify the information for the targeted mulci-cloud infrastructure
          (MCI)
        in: body
        name: mciInfo
        required: true
        schema:
          $ref: '#/definitions/controller.MigrateInfraRequest'
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated to the multi-cloud infrastructure
          schema:
            $ref: '#/definitions/controller.MigrateInfraResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Migrate an infrastructure to the multi-cloud infrastructure (MCI) with
        defaults
      tags:
      - '[Migration] Infrastructure'
  /migration/ns/{nsId}/mci/{mciId}:
    delete:
      consumes:
      - application/json
      description: Delete the migrated mult-cloud infrastructure (MCI)
      operationId: DeleteInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mmci01
        description: Migrated Multi-Cloud Infrastructure (MCI) ID
        in: path
        name: mciId
        required: true
        type: string
      - default: terminate
        description: Action for deletion
        enum:
        - terminate
        - force
        in: query
        name: action
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The result of deleting the migrated multi-cloud infrastructure
            (MCI)
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Delete the migrated mult-cloud infrastructure (MCI)
      tags:
      - '[Migration] Infrastructure'
    get:
      consumes:
      - application/json
      description: Get the migrated multi-cloud infrastructure (MCI)
      operationId: GetInfra
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mmci01
        description: Migrated Multi-Cloud Infrastructure (MCI) ID
        in: path
        name: mciId
        required: true
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The migrated multi-cloud infrastructure (MCI) information
          schema:
            $ref: '#/definitions/controller.MigrateInfraResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Get the migrated multi-cloud infrastructure (MCI)
      tags:
      - '[Migration] Infrastructure'
  /migration/ns/{nsId}/mciWithDefaults:
    post:
      consumes:
      - application/json
      description: Migrate an infrastructure to the multi-cloud infrastructure (MCI)
        with defaults.
      operationId: MigrateInfraWithDefaults
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Specify the information for the targeted mulci-cloud infrastructure
          (MCI)
        in: body
        name: mciInfo
        required: true
        schema:
          $ref: '#/definitions/controller.MigrateInfraWithDefaultsRequest'
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated to the multi-cloud infrastructure
          schema:
            $ref: '#/definitions/controller.MigrateInfraWithDefaultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-beetle_pkg_api_rest_model_beetle.Response'
      summary: Migrate an infrastructure to the multi-cloud infrastructure (MCI) with
        defaults
      tags:
      - '[Migration] Infrastructure'
  /migration/ns/{nsId}/resources/securityGroup:
    get:
      consumes:
      - application/json
      description: Get the list of all migrated security groups in the namespace
      operationId: ListMigratedSecurityGroups
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Option
        enum:
        - id
        in: query
        name: option
        type: string
      - description: 'Field key for filtering (ex: systemLabel)'
        in: query
        name: filterKey
        type: string
      - description: 'Field value for filtering (ex: Registered from CSP resource)'
        in: query
        name: filterVal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/controller.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/resource.RestGetAllSecurityGroupResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all migrated security groups
      tags:
      - '[Migration] Resources for VM infrastructure'
    post:
      consumes:
      - application/json
      description: Create a new migrated security group in the namespace
      operationId: CreateMigratedSecurityGroup
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: 'Option: [required params for register] connectionName, name,
          vNetId, cspResourceId'
        enum:
        - register
        in: query
        name: option
        type: string
      - description: Details for an securityGroup object
        in: body
        name: securityGroupReq
        required: true
        schema:
          $ref: '#/definitions/model.TbSecurityGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSecurityGroupInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create a migrated security group
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/securityGroup/{securityGroupId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific migrated security group in the namespace
      operationId: DeleteMigratedSecurityGroup
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Security Group ID
        in: path
        name: securityGroupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete a migrated security group
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/securityGroup/{sgId}:
    get:
      consumes:
      - application/json
      description: Get details of a specific migrated security group in the namespace
      operationId: GetMigratedSecurityGroup
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Security Group ID
        in: path
        name: sgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSecurityGroupInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get a specific migrated security group
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/vNet:
    get:
      consumes:
      - application/json
      description: Get the list of all migrated virtual networks in the namespace
      operationId: ListMigratedVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/controller.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/resource.RestGetAllVNetResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all migrated virtual networks
      tags:
      - '[Migration] Resources for VM infrastructure'
    post:
      consumes:
      - application/json
      description: Create a new migrated virtual network in the namespace
      operationId: CreateVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Virtual Network creation request
        in: body
        name: vNetReq
        required: true
        schema:
          $ref: '#/definitions/model.TbVNetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbVNetInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create a migrated virtual network
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/vNet/{vNetId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific migrated virtual network in the namespace
      operationId: DeleteMigratedVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Virtual Network ID
        in: path
        name: vNetId
        required: true
        type: string
      - description: Action
        enum:
        - withsubnets
        - refine
        - force
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete a migrated virtual network
      tags:
      - '[Migration] Resources for VM infrastructure'
    get:
      consumes:
      - application/json
      description: Get details of a specific virtual network in the namespace
      operationId: GetMigratedVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Virtual Network ID
        in: path
        name: vNetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbVNetInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get a specific migrated virtual network
      tags:
      - '[Migration] Resources for VM infrastructure'
...
  /recommendation/containerInfra:
    post:
      consumes:
      - application/json
      description: |-
        Recommend an appropriate container infrastructure for container-based workloads

        [Note] `desiredProvider` and `desiredRegion` are required.
        - `desiredProvider` and `desiredRegion` can be set in the query parameter or the request body.
        - If both are set, the values in the request body take precedence.
      operationId: RecommendContainerInfra
      parameters:
      - description: Specify the source container infrastructure
        in: body
        name: UserInfra
        required: true
        schema:
          $ref: '#/definitions/controller.RecommendInfraRequest'
      - default: aws
        description: Provider (e.g., aws, azure, gcp)
        enum:
        - aws
        - azure
        - gcp
        - ncp
        in: query
        name: desiredProvider
        type: string
      - default: ap-northeast-2
        description: Region (e.g., ap-northeast-2)
        in: query
        name: desiredRegion
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The result of recommended container infrastructure
          schema:
            $ref: '#/definitions/controller.RecommendInfraResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SimpleMsg'
      summary: Recommend an appropriate container infrastructure for cloud migration
      tags:
      - '[Recommendation] Infrastructure'
  /recommendation/mci:
    post:
      consumes:
      - application/json
      description: |-
        Recommend an appropriate VM infrastructure (i.e., MCI, multi-cloud infrastructure) for cloud migration

        [Note] `desiredCsp` and `desiredRegion` are required.
        - `desiredCsp` and `desiredRegion` can set on the query parameter or the request body.

        - If desiredCsp and desiredRegion are set on request body, the values in the query parameter will be ignored.
      operationId: RecommendVMInfra
      parameters:
      - description: Specify the your infrastructure to be migrated
        in: body
        name: UserInfra
        required: true
        schema:
          $ref: '#/definitions/controller.RecommendVmInfraRequest'
      - default: aws
        description: Provider (e.g., aws, azure, gcp)
        enum:
        - aws
        - azure
        - gcp
        - ncp
        in: query
        name: desiredCsp
        type: string
      - default: ap-northeast-2
        description: Region (e.g., ap-northeast-2)
        in: query
        name: desiredRegion
        type: string
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The result of recommended infrastructure
          schema:
            $ref: '#/definitions/controller.RecommendVmInfraResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SimpleMsg'
      summary: Recommend an appropriate VM infrastructure (i.e., MCI, multi-cloud
        infrastructure) for cloud migration
      tags:
      - '[Recommendation] Infrastructure'
...
  /test/tracing:
    get:
      consumes:
      - application/json
      description: Test tracing to Tumblebug
      operationId: TestTracing
      parameters:
      - description: 'Custom request ID (NOTE: It will be used as a trace ID.)'
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SimpleMessage'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.SimpleMessage'
      summary: Test tracing to Tumblebug
      tags:
      - '[Test] Utility'
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"